+++ scripty_botlists/src/lists/botlist_me/models.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Serialize)]
 pub struct PostStats {
 	pub server_count: usize,
-	pub shard_count:  u32,
+	pub shard_count:  u16,
 }
 
 #[derive(Debug, Deserialize, Copy, Clone)]
