+++ scripty_botlists/src/lists/discord_bots_gg/models.rs
@@ -2,7 +2,7 @@
 #[serde(rename_all = "camelCase")]
 pub struct PostStats {
 	pub server_count: usize,
-	pub shard_count:  u32,
+	pub shard_count:  u16,
 }
 
 #[derive(Deserialize, Debug, Clone)]
