+++ scripty_botlists/src/lists/topcord_xyz/models.rs
@@ -1,7 +1,7 @@
 #[derive(Serialize, Debug, Copy, Clone)]
 pub struct PostStats {
 	pub servers: usize,
-	pub shards:  u32,
+	pub shards:  u16,
 }
 
 #[derive(Deserialize, Debug, Clone)]
