+++ scripty_botlists/src/lists/voidbots_net/models.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Serialize)]
 pub struct PostStats {
 	pub server_count: usize,
-	pub shard_count:  u32,
+	pub shard_count:  u16,
 }
