+++ scripty_webserver/src/endpoints/bot_stats/mod.rs
@@ -19,7 +19,7 @@ pub struct BotStats {
 	pub guild_count:   usize,
 	pub user_count:    usize,
 	pub channel_count: usize,
-	pub shard_count:   u32,
+	pub shard_count:   u16,
 }
 
 pub async fn get_bot_stats() -> Result<Json<BotStats>, WebServerError> {
